Cover tests.
